Move ring_buffer
