+++ src/library/utilities/ring_buffer.cpp
